+++ app/lib/constants.js
@@ -1,4 +1,25 @@
+/* eslint-disable sort-keys */
 module.exports = {
   ASSETS_URL: 'https://assets.nhs.uk',
   SITE_ROOT: '/find-a-chlamydia-test',
+  SYMPTOMS: {
+    yes: 'yes',
+    no: 'no',
+  },
+  AGE: {
+    under16: '1',
+    '16to25': '2',
+    over25: '3',
+  },
+  SERVICE_CHOICES: {
+    symptoms: '0',
+    under16: '1',
+    '16to25': '2',
+    over25: '3',
+  },
+  SERVICE_TYPES: {
+    professional: 'professional',
+    kit: 'kit',
+  },
 };
+/* eslint-enable sort-keys */