+++ config/routes.js
@@ -1,4 +1,7 @@
 const router = require('express').Router();
+const locationValidator = require('../app/middleware/locationValidator');
+const notInEnglandHandler = require('../app/middleware/notInEnglandHandler');
+const postcodeLookup = require('../app/middleware/postcodeLookup');
 const renderer = require('../app/middleware/renderer');
 const selectionValidatorSymptoms = require('../app/middleware/selectionValidatorSymptoms');
 const selectionValidatorAge = require('../app/middleware/selectionValidatorAge');
@@ -28,6 +31,19 @@ router.get(
   '/choose',
   selectionValidatorAge,
   renderer.choose
+);
+
+router.get(
+  '/location',
+  renderer.location
+);
+
+router.get(
+  '/results',
+  locationValidator,
+  postcodeLookup,
+  notInEnglandHandler,
+  renderer.results
 );
 
 module.exports = router;